Use `psycopg` in `scripts/import_data.py`
The `scripts/import_data.py` failed because `psycopg2` was not
installed. So switch it to use `pscyopg`, which is v3, and update the
connection parameters to a single connection string.

Update the README to reflect this change, and add language IDs to the
fenced code blocks. Also switch from `wget` to `curl`.