Add FDW selector that queries a separate schema
On the assumption that the ClickHouse foreign tables would be imported
into a schema named `fdw`, add an additional option to the database
selector, "fdw". The display needs some CSS tweaking and it needs a
custom icon, but it does work.

Change the selector code to set the value to "fdw" when appropriate.

In `runQuery()`, replace `FROM uk` with `FROM fdw.uk` in the query when
the database choice is "fdw", and change `getQuery()` to also recognize
"fdw" as an option.