Add Postgres FDW Large queries
Currently dupes the regular Postgres queries but replaces the some of
the table names to use the ClickHouse materialized views.

Update the `SelectArea()` function to enable the medium/large buttons
for both FDW and CLickHouse, and update `getQuery()` to fetch the large
FDW queries when wanted.

Add DDL to `schemas/postgres.sql` to create the `clickhouse_fdw`
extension and import the tables from the demo database so that they're
accessible to HouseClick. This won't work without `clickhouse_fdw`, so
add a placeholder to `README.md` for forthcoming installation
instructions.

While at it, expand and reorder the configs in `.env.example` and`
remove some trailing spaces from files and restore the `EXTRACT()`
function to the Postgres queries, since it was fixed by
ClickHouse/clickhouse_fdw#31.